fix serious data corruption bug Many of the MetaData concepts had been defined without calling setPrimordialUUID.  This caused the builder to generate a random UUID for each concept.  MetaData concepts CANNOT BE RANDOM.  Everything would work for one run through building ibdf files / databases / running the system.  But then would randomly fail later, with missing concept refs to the constants - any time that the MetaData generation code was re-run - as new random UUIDs would be assigned.
This fix changes ALL MetaData concepts to be properly specified with
Type-5 UUIDs (Based on their FSN) if no UUID was assigned.  This also
cleans up some formatting issues in the IsaacMetaData class that made
debugging it difficult, reorganized a couple of metadata conepts that
were not in the right places, and replaced the hard-coded class refs
with HK2 refs in the ExportTerminology mojo.