getting rid of a bunch of IBDF diff'ing code that was in a bad state when brought over here.
The design of the diff was completely broken from the beginning, it
wasn't thread safe, it held everything in memory multiple times, and was
extremely inefficient.

While there may be a branch of this code off in VA-land that someone
picked up and actually fixed.... it is 10 times easier to start from
scratch at this point.