Created isolated versions of graph collector and supporting classs.
The stamp coordinates themselves where getting modified as the windows where opening, and the lack of isolation was causing concepts to have incorrect/missing parents computed. This was showing up as multiple roots.

The stamp coordinate is cloned, and readonly inside of a relative version computer, and this seems to resolve the multiple roots problems.