Misc fixes for issues identified importing change sets.
Handle notification of changes with a commit record, only import new versions if more that a time is different. Make sure that classified content does not create duplicate semantics for a single component when classified in two different environments, and then later synchronized. Modify change set loader to track size of change set files, not just a boolean if imported or not. Allow the change set importer to make use of the commit service commit record notification.